app/src/main/java/com/ilhmdhn/movieapps/ui/home/tvshow/TvShowViewModel.kt
package com.ilhmdhn.movieapps.ui.home.tvshow

import androidx.lifecycle.LiveData
import androidx.lifecycle.ViewModel
import androidx.paging.PagedList
import com.ilhmdhn.movieapps.data.MovieRepository
import com.ilhmdhn.movieapps.data.source.local.entity.ListTvShowEntity
import com.ilhmdhn.movieapps.vo.Resource

class TvShowViewModel(private val movieRepository: MovieRepository) : ViewModel() {

    fun getTvShow(): LiveData<Resource<PagedList<ListTvShowEntity>>> = movieRepository.getAllTv()
}